Fixed a slight type issues in bitwise.h
Fixed a slight type issues in bitwise.h, determined that at some point, SHA256_M() segfaults, however it has not been determined where it segfaults.

+++ include/crypt/bitwise.h
@@ -3,12 +3,12 @@
 
 #include <stdint.H>
 
-static uint16_t rotl8(uint8_t rot, uint8_t by)
+static uint8_t rotl8(uint8_t rot, uint8_t by)
 {
     return (rot << by) | (rot >> (16 - by));
 }
 
-static uint16_t rotr8(uint8_t rot, uint8_t by)
+static uint8_t rotr8(uint8_t rot, uint8_t by)
 {
     return (rot >> by) | (rot << (8 - by));
 }
